Fix Decode (float64 en lugar de float32)

+++ go/dto/Camion.go
@@ -8,7 +8,7 @@ import (
 type Camion struct {
 	Patente                  string    `json:"patente"`
 	PesoMaximo               int       `json:"peso_maximo"`
-	CostoPorKilometro        float32   `json:"costo_por_kilometro"`
+	CostoPorKilometro        float64   `json:"costo_por_kilometro"`
 	FechaCreacion            time.Time `json:"fecha_creacion"`
 	FechaUltimaActualizacion time.Time `json:"fecha_ultima_actualizacion"`
 	IdCreador                string    `json:"id_creador"`
