+++ go/dto/ProductoPedido.go
@@ -9,8 +9,8 @@ type ProductoPedido struct {
 	CodigoProducto string  `json:"codigo_producto"`
 	Nombre         string  `json:"nombre_producto"`
 	Cantidad       int     `json:"cantidad"`
-	PrecioUnitario float32 `json:"precio_unitario"`
-	PesoUnitario   float32 `json:"peso_unitario"`
+	PrecioUnitario float64 `json:"precio_unitario"`
+	PesoUnitario   float64 `json:"peso_unitario"`
 }
 
 // Metodo que sirve para crear un ProductoPedido para un pedido
