+++ go/model/Camion.go
@@ -10,7 +10,7 @@ type Camion struct {
 	ObjectId                 primitive.ObjectID `bson:"_id,omitempty"`
 	Patente                  string             `bson:"patente"`
 	PesoMaximo               int                `bson:"peso_maximo"`
-	CostoPorKilometro        float32            `bson:"costo_por_kilometro"`
+	CostoPorKilometro        float64            `bson:"costo_por_kilometro"`
 	FechaCreacion            time.Time          `bson:"fecha_creacion"`
 	FechaUltimaActualizacion time.Time          `bson:"fecha_ultima_actualizacion"`
 	IdCreador                string             `bson:"id_creador"`
